backend/requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.13,<4.0
django-cors-headers>=3.8,<4.0
psycopg2>=2.9,<3.0
